// src/enums/common.enum.ts
export enum StatusColor {
    Primary = 'primary',
    Info = 'info',
    Success = 'success',
    Warning = 'warning',
    Danger = 'danger',
}

export enum RespStatusCode {
    Ok = 9999,
    Error = -9999,
}

export enum Gender {
    Male = 'MALE',
    Female = 'FEMALE',
    Other = 'OTHER',
}

export enum Percent {
    Zero = 0,
    OneHundred = 100,
}

export enum CostType {
    Percent = 'PERCENT',
    Money = 'MONEY',
}

export enum PlatformType {
    WebApp = 'WEB_APP',
    CustomerApp = 'CUSTOMER_APP',
    CounterApp = 'COUNTER_APP',
}

export enum PaymentMethod {
    Cash = 'CASH',
    Card = 'CARD',
    Giftcard = 'GIFTCARD',
    Reward = 'REWARD',
    SplitPayment = 'SPLIT_PAYMENT',
    Other = 'OTHER',
}

export enum DayOfWeekType {
    Monday = 'MONDAY',
    Tuesday = 'TUESDAY',
    Wednesday = 'WEDNESDAY',
    Thursday = 'THURSDAY',
    Friday = 'FRIDAY',
    Saturday = 'SATURDAY',
    Sunday = 'SUNDAY',
}

export enum ArrowDirection {
    Left = 'left',
    Right = 'right',
    Up = 'up',
    Down = 'down',
}