// src/components/common/avatar/CAvatar.type.ts
export enum AvatarSize {
    /**
     * @constant `16`
     */
    Tiny = 'tiny',
    /**
     * @constant `20`
     */
    Small = 'small',
    /**
     * @constant `32`
     */
    SmallMedium = 'smallMedium',
    /**
     * @constant `40`
     */
    Medium = 'medium',
    /**
     * @constant `48`
     */
    MediumLarge = 'mediumLarge',
    /**
     * @constant `56`
     */
    Large = 'large',
    /**
     * @constant `64`
     */
    LargeExtra = 'largeExtra',
    /**
     * @constant `72`
     */
    ExtraLarge = 'extraLarge',
    /**
     * @constant `96`
     */
    Extra = 'extra',
    /**
     * @constant `128`
     */
    ExtraExtra = 'extraExtra',
}